TrackStore klasse tilføjet, tracks kan nu gemmes!!!

+++ ATM.Application/ATM.ClassLib/TrackObject.cs
@@ -14,13 +14,13 @@ namespace ATM.ClassLib
             XCoord = Int32.Parse(arr[1]);
             YCoord = Int32.Parse(arr[2]);
             Altitude = Int32.Parse(arr[3]);
-            TimeStamp = Int32.Parse(arr[4]);
+            TimeStamp = long.Parse(arr[4]);
         }
 
         public string Tag { get; set; }
         public int XCoord { get; set; }
         public int YCoord { get; set; }
         public int Altitude { get; set; }
-        public int TimeStamp { get; set; }
+        public long  TimeStamp { get; set; }
     }
 }
